+++ common/src/main/java/fr/loudo/narrativecraft/narrative/story/StoryHandler.java
@@ -551,6 +551,7 @@ public final class StoryHandler implements InkTagHandler.Lifecycle, IStoryHandle
         characterEntities.clear();
         inkTagHandler.stopAll();
         playerSession.closeEditor();
+        playerSession.clearInteractionSession();
         dialogVisible = false;
         choicesVisible = false;
     }